backend/app/pyproject.toml
[tool.poetry]
name = "digitalvixen"
version = "0.1.0"
description = ""
authors = ["darklorian <lolpokens@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.9"
psycopg-binary = "^3.1.9"
djangorestframework-simplejwt = "^5.2.2"
httpx = "^0.24.1"
django-minio-backend = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

